% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx.R
\name{cdx}
\alias{cdx}
\title{Changes directory based on an external reference document}
\usage{
cdx(targ = "list")
}
\arguments{
\item{targ}{A projectory reference id.}
}
\description{
Changes directory based on an external reference document
}
\examples{
cdx("NILM")
cdx("rbsaSM2")


}
