\name{EcotopePackage-package}
\alias{EcotopePackage-package}
\alias{EcotopePackage}
\docType{package}
\title{
Various utility codes for internal Ecotope Use
}
\description{
All the utility functions
}
\details{
\tabular{ll}{
Package: \tab EcotopePackage\cr
Type: \tab Package\cr
Version: \tab 0.0.0.9000\cr
Date: \tab 2015-04-01\cr
License: \tab GPL-2\cr
}

}
\author{
Jeffrey Uslan

Maintainer: Jeffrey Uslan, jeffreyu@ecotope.com

}
\references{
References to come later
}

\examples{


}

