% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.seem.R
\name{read.seem}
\alias{read.seem}
\title{Helper function to read SEEM outputs into R}
\usage{
read.seem(seem_output)
}
\arguments{
\item{seem_output}{An output CSV generated by SEEM.}
}
\description{
Helper function to read SEEM outputs into R
}
\examples{
read.seem("seem_outputs.csv")


}
