% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/stata_time_to_R_time.R
\name{stata_time_to_R_time}
\alias{stata_time_to_R_time}
\title{Convert Stata time to R time after using read.dta}
\usage{
stata_time_to_R_time(times)
}
\arguments{
\item{times}{A time vector to convert.}
}
\description{
Convert Stata time to R time after using read.dta
}

